Implement intrusive reference-counted pointer type
See the docs on the pointer type for all the details but the basic
idea is that the pointee type has a refcount member and that is
used by ref to keep track of everything.

It also has several utilities for converting existing pointers into
references and allows casting non-const references to const
references transparently as long as the refcount member is mutable.