Add default implementation for Model::evaluate
It doesn't make sense for every single derived model type to
implement this when it can be implemented purely using the lookup
method of the Model interface. Beyond that, the implementation is
actually somewhat complicated so it's best that there is a common
implementation.

Certain model types may still want to specialize this and for that
I've left it virtual.

In addition, now that there's the Value type it makes far more
sense for Model::evaluate and Model::lookup to return a value than
it does for them to return an Operation reference that is only
constrained to ConstantInt/ConstantFloat by comment specs. As such,
I've changed them to use the Value type.