Make Z3Solver work with anonymous constants
Constants created as part of the allocation process don't have names
and instead are denoted by numbers. Z3Solver didn't support that so
this PR adds support for it.

I've also added some logic to simplify the generated Z3 AST by avoiding
execessive conversions between bitvectors and booleans in cases where
we have a good guess of what they'll be used for.

Beyond that, this moves some code around to avoid include z3++.h in
public header files. It is very expensive to compile so I figure the
indirection introduced as part of doing this is worth it.