Test MemHeap and fix related bugs (#48)
* Simplify include paths

* Add pointer resolve test case

* Make Z3Solver work with anonymous constants

Constants created as part of the allocation process don't have names
and instead are denoted by numbers. Z3Solver didn't support that so
this PR adds support for it.

I've also added some logic to simplify the generated Z3 AST by avoiding
execessive conversions between bitvectors and booleans in cases where
we have a good guess of what they'll be used for.

Beyond that, this moves some code around to avoid include z3++.h in
public header files. It is very expensive to compile so I figure the
indirection introduced as part of doing this is worth it.

* Rename unit tests to start with unit/

* Make MemHeap pass initial test case

Still probably lots of work to do here but it's much more likely to
be correct now. This also fixes up several generated assertions
which were incorrect.

* Fix normalize_to_bv