Give failure logger required context to see all assertions (#64)
The failure logger was not printing certain constants because it was not being given access to the assertion that was passed in as part of the check call to the solver and therefore couldn't see those constants. This fixes that by adding an extra argument to the FailureLogger class containing the required assertion.