Simplify Operation internals
Since Operation now has a constructor that directly takes the
variant type there are several constructors that can just be deleted.

In addition, it turns out that we actually need to define our own
copy/move ctor/assignment operators as refcount should not be modified
during a copy or move.