Simplify Value implementation
This simplifies the value implementation by replacing a manual union
with a variant and then rewriting the implementation to work with that.

I've also taken the opportunity to remove redundant assertions now that
the arithmetic methods have switched to using the proper accessors.
(e.g. apint() already asserts on is_int(), there's no need to repeat
that within the body of add)