Keep pointers resolved for constant getelementptr operations (#156)
For a number of yarpgen tests the majority of the time is spent
resolving pointers that don't already have an associated allocation.
This was made worse by the fact that constant getelementptr instructions
were stripping out the attached allocation. This is just a point fix to
keep the corresponding allocation.