Don't catch exceptions when running tests individually
If gtest catches exceptions thrown from tests then we don't get to see
backtraces showing where the exception was thrown. This is normally a
good thing since it allows up to keep running tests in the rest of the
test suite, however when running the tests from ctest we run each one
individually so if an uncaught exception is thrown we want to just die
and print out the backtrace.