Add method to print an expression while in a debugger
This is a private method which prints a representation of the expression
to stdout. It's meant to be used as an aid while debugging so I've left
it as a private method.

Note that this only really works when the caffeine library is built as a
shared library as otherwise the linker will delete the method since it
is unused.