Fix use of CopyVTable within Operation (#212)
As it turns out, the copy constructor for CopyVTable runs before the
derived class actually initializes the derived vtable. This means that
as soon as we actually start copying Operations around the vtable was
not being copied around with them.

The fix for this is to have Operation call copy_vtable within its ctors.
I've prevented the same issue from (hopefully) occurring again by
deleting the copy/move ctors of CopyVTable and documenting this footgun
in the class documentation.