Refactor constant folding out into a separate visitor (#213)
* Refactor constant folding code out of CreateXXX methods

This moves the constant folding code into a visitor class within
src/IR/Operation.h. As a result a number of methods can be simplified
back into macro definitions and I'm able to delete a large amount of
code within Operation.cpp.

The motivation for this was the fact that with_new_operands doesn't
correctly perform constant folding as we'd expect it to (in fact it
doesn't perform it at all). In order to run constant folding on an
arbitrary expression we need to be able to run it on arbitrary operation
nodes and therefore here we are.

* Apply constant folding within with_new_operands