Split format target into multiple sub-commands
This allows it to run faster when you've only changed a few files.

I don't have hard numbers but previously running a format always took on
the order of about 10 seconds for me but now it takes ~2s even with the
yarpgen files that I have in my repo (which are usually slower to
format).