Optimize cmake configure times
The main contributors that I found to cmake configure being slow were
two different non-required calls to find_package that didn't succeed.
These took on the order of a few hundred milliseconds which noticably
slowed down the configure step.

To remedy this I cached the failure value in cases where the
find_package calls failed but we were able to successfully find the
library somewhere else. This seems to have halved the time we spend on
running cmake configures when I try it locally.