Rewrite cmake utilities for building LLVM IR libraries (#273)
* Including BitcodeWriter.h in CI causes issues, remove it

* Fix include directories for gen-test-main target

* Rewrite LLVM library cmake utility

* Remove unused lookup_const calls from caffeine.cpp

This broke once the entry point was changed

* Add support for main as an entry point

In this case the arguments are not meant to be symbolic so we should
just default them to valid values.

Specifically, I've chosen to default to
  argc = 0
  argv = nullptr

* Rewrite cmake methods for creating llvm targets

We no longer use a custom language to make everything work. Now it is
all done by a sequence of custom commands and targets. As a side effect,
this should make it possible to compile on MSVC again.

* Remove some unused logging

* Add support for ninja depfiles when supported