Redesign context forking API (#279)
The fork API for context is rather inconvenient to work with. This PR changes it to return a vector of contexts so that it's not necessary distinguish between the current context and all the forks.

The old API is still available as Context::fork_once when that is more convenient.