Fix bad dependencies for LLVM_C[XX] files (#287)
In the previous setup clang would generate a depfile with the
dependencies specified using absolute paths. However, ninja expects
the depfile target path to be relative to the binary directory. To
fix this we need to update both the WORKING_DIRECTORY and the output
filename for the compilation commands.

As an extra bonus, this also sets a property to disable useless
"Built target X" message when building using makefiles.