Introduce tool to generate implementations of LLVM builtins
The current state of the repo (as of this commit) is that we have
builtin implementations for the most common calls to the llvm.memset.*
intrinsic. Namely, we really only support the
  void llvm.memset.p0i8.i64(i8* %dst, i8 %val, i64 %len, i1 %isvolatile)
variant. This is done by hand-implementing a function
  void caffeine_builtin_memset(...)
with the same signature in _LLVM IR_ and then having the interpreter
patch calls to the intrinsic to instead point to the function we
implemented by hand.

This has some big limitations:
1. We don't support anything other than the most basic memset calls.
   Anything that gets generated to use a different type will fail to
   compile. As an example, this will fail on platforms where pointers
   are 32-bits in size.
2. We don't support any platform that is not data-layout-compatible with
   x86_64-linux-gnu. This is the more relevant issue because it means
   that we cannot compile our test cases for windows.

Both limitations are decently large problems. The solution I've
undertaken here is to instead generate the required builtin
implementations at build time. This means that we never end up needing
an unsupported function variant since we support _all_ the variants.

As a bonus, we can now also use IR names that can't be defined from C.