Break solver out of Context (#299)
This PR breaks out the `Solver` from the `Context`. This is helpful in order to allow multiple `Solver`s to coexist in a multithreaded environment. As part of this change, the `Interpreter` now creates the `Solver` stack, as opposed to the caffeine binary harness. This is to allow each `Interpreter` to own it's `Solver` so that we can have a `Solver` per thread. Previously, it was only possible to have one `Solver` per runtime.

I was considering adding a method to `Context` along the lines of `setSolver` but decided against it because of the ownership issue.