Rewrite deps handling to not include dependency as a subdirectory
At the moment, FetchContent will always add the fetched depenency as a
subdirectory. This works ok if the dependency's CMakeLists is written to
work properly when added as a subdirectory. Unfortunately, the CMake
configurations for most projects are... not. This limits the number of
dependencies that we can bundle using FetchContent to only a few.

However, cmake does already have support for doing all the difficult
bits of bundling dependencies by using ExternalProject_Add.
Unfortunately this doesn't help us directly as ExternalProject_Add is
targetted towards building a super-project that assembles multiple
subprojects into one big build.

This can be worked around by generating a simple cmake project that does
nothing but call ExternalProject_Add. This is also effectively what the
FetchContent module does except that it disables all the steps except
the download one and expects you to use add_subdirectory.

So to include a dependency we do the following:
1. First, check to see if we can find the project via find_package. If
   so then we're done.
2. Generate a CMakeLists.txt in a project-specific subdirectory of the
   _deps folder in the binary dir. We pick the name so as not to
   conflict with the directory names used by FetchContent.
3. Configure the subproject to install to a subdirectory within _deps.
   I've chosen to use install for this. If we add a project named
   install as a depedency then we'll have to change this. I don't think
   this will be an issue.
4. Build the project (during caffeine configuration).
5. And finally, use find_package to find the newly installed files
   within the local installation directory.

Once the project is built we'll stop after step 1 since we find the
installed files from the last time we ran configure.

Note that there is some extra work that happens to cover these cases:
- Multi-configuration generators (e.g. Visual Studio, XCode): To make
  these work we need to build the subproject for all supported
  configurations.
- Updating the config file: Need to always rebuild the subproject if the
  config has changed.