Remove all non-const methods from Operation
Now that everything is being handled through OpRefs (which are const)
these non-const methods are no longer useful. Moreover, they constrain
possible implementations so it's better just to get rid of them.

This also rewrites the operand iterator to not assume that the
underlying representation of operands is linear in memory.