Make ExprEvaluator::visitSelectInst work with pointers
Previously this would de-pointerize pointers when used to select between
two pointers, resulting in an integer with the same bitwidth as a
pointer. This would cause any subsequent pointer operation using that
pointer to fail.

This commit changes this to return a pointer object if the operands are
pointers.

Fixes #359