Fix zext Z3Solver method for 1bit integers
The problem was that Z3Solver::visitZExt was trying to get the bitwidth
out of the sort of the argument before the argument had been normalized
to a bitvector type. When the input was a boolean (e.g. from a
comparison) it obviously doesn't have a bitwidth and this resulted in a
crash. The fix is just to use the sort of the expr after it's been
normalized to a bitvector.

Fixes #361