Implement non-symbolic memory allocation (#372)
This commit implements a basic concrete allocator. This avoids the
problem where creating new allocations results in SAT queries increasing
in size quadratically due to the constraints about queries not
overlapping. Since the assertions related to allocations are now all
constant we avoid creating any new assertions as they would all be
constant.

The implementation of the allocator used here is a pretty basic external
buddy allocator. For an explanation of buddy allocation see [1]. The
main difference to this one is that it is built using persistent data
structures so that we can efficiently share the common parts of the heap
between different contexts.

At the moment the heap does not support re-merging blocks that have been
split. However, if allocation ever fails then it falls back to fully
symbolic allocation so this shouldn't actually limit anything in that
case. The same happens if an allocation with a symbolic size occurs.
Ideally, this should happen infrequently as the fully-symbolic
allocation mode is quite inefficient.

[1]: https://en.wikipedia.org/wiki/Buddy_memory_allocation

Squashed commit messages:
* Implement basic concrete memory allocator
* Allow alignments of 0
* Allow configuring whether MemHeapMgr creates concrete heaps
* Add command-line option to disable the concrete allocator