Add config header and CAFFEINE_ENABLE_TRACING option
This commit changes caffeine to generate a header caffeine/Config.h.
Currently, the only thing that it contains is a definition reflecting
the value of CAFFEINE_ENABLE_TRACING in the cmake configuration step.

The config is placed in ${CMAKE_BUILD_DIR}/gen/caffeine just like all
the other existing generated files.