Implement basic tracing framework
This commit adds a fairly simple tracing library. It is based off of
the APIs of the some of the other tracing libraries I've seen,
particularly Artillery within FB. In this case, the feature set is much
more limited. In fact, the only thing supported is to log a span (a
named period of time) which can then have arbitrary annotations attached
to it.

All the design here does with the spans is to log them to a trace file.
The format of this file is just repeated serialized TraceSpan messages
which are appended to the file as spans are completed. The current
design just uses fwrite to write the entire message at once. Since
fwrite is specified to be atomic on both windows and any POSIX-compliant
system this should be safe for our use-cases.

For my needs (seeing which operations are slow) this is good enough. To
ensure that it doesn't affect performance I've made sure that all the
methods on TraceContext and AutoTraceBlock are replaced with stubs when
CAFFEINE_ENABLE_TRACING is not set.