Optimize constant-size symbolic arrays (#381)
* Allow extracting the corresponding llvm::fltSemantics from a Type

* Fix memory error within MemHeap::operator[]

* Allow Solver::evaluate to make up values in certain cases

When a constant isn't present in the model then it doesn't matter what
value it takes. In this case we just return a 0 value.

This commit only implements this for integers and floating point types.
This is all that is needed to make the current feature work.

* Represent symbolic arrays as a FixedSizeArray of constants

This allows each byte of the constant array to be managed separately by
the underlying solver which makes solver calls much faster.