Fix assert failure for vector GetElementPtr instructions
Previously this code assumed that if the GEP instruction instance was a
vector instruction then the pointer operand would be a vector as well.
However, this isn't guaranteed and we crash if some of the indices are
vectors and the pointer operand is not.

This commit fixes the vector broadcasting to be based on the final type.