Add simplifications to optimize away load-store operations
When a value is stored and then loaded again we cut it into bytes and
then put it back together again. This doesn't change the acutal value
but it does result in the expression getting larger every time. The
optimization rules added in this commit reduce this down to a no-op when
applicable.