Avoid performing a solver call if the result is obvious
If any of the assertions is a literal false then we can tell that
the result of the solver call will be UNSAT without having to actually
build a solver to query. Since building the solver takes a significant
amount of time this can result in some rather large speedups.