Various optimizations to pointer resolution (#461)
* Rewrite Interpreter::visitLoad to use TransformBuilder

* Rewrite Interpreter::visitStore to use TransformBuilder

* Avoid going through the context store when there's only a single context

* Avoid performing a solver call if the result is obvious

If any of the assertions is a literal false then we can tell that
the result of the solver call will be UNSAT without having to actually
build a solver to query. Since building the solver takes a significant
amount of time this can result in some rather large speedups.