Implement better multithreaded context store (#464)
Currently we use QueueingContextStore (or a variant thereof that caches the last value). However, this has a problem that if there isn't enough work to saturate all the worker threads then the unused threads will spend a bunch of their time bouncing through locks when new values are inserted into the queue. This results in caffeine using far more CPU than is useful. Furthermore, when new values are inserted generally a different thread will win the race to pull the context out of the store. This means that there is little cache locality when dealing with these contexts.

This PR introduces a new context store called ThreadQueueContextStore that keeps a separate deque for each thread. When a thread needs to get a new context it'll pull the most recent one off of it's own queue. If it has no work items then it will pull the oldest element of the queue of a random thread. If unable to find an element in the queue then it will sleep for a period before trying again. This ensures that inactive threads use little CPU while also ensuring that they find work in a reasonable amount of time.

This PR also includes a change to ThreadMap so that it can be used with non-default-initializable types as well as some small cleanups within the Tracing.h. Furthermore, I've deleted the old ThreadQueuedExecutor since it is worse in every single use case compared to the new one.

Closes #459