Refactor solver construction and inheritance (#465)
This commit changes solvers to be a proper solver stack, removing the
SequenceSolver and instead replacing it with the new TransformSolver
base class. These changes then allow for writing a general solver
builder class (SolverBuilder) that can build a solver stack on demand
from a sequence of user-provided functions.

With that in place, it changes all the existing spots where solvers were
constructed to instead use the builder class. There is now one place to
build the default solver stack and, for now, that is used everywhere a
SolverBuilder instance is created.