`Invoke` instruction implementation (#480)
* Implement `Invoke` inst

The implementation of `Invoke` is the same as `Call` because we're
branching to a new function in the same way. The only difference is
in returning from the function, which will be handled by the
`resume` instruction.

* Implement `Invoke` return semantics

Although the `Invoke` instruction itself was implemented in a previous
commit, the `Return` instruction did not account for the fact that it
had to use a different control flow scheme if the caller was an `Invoke`
instruction. This commit fixes that

* Remove logging and add libc to coverage

* Actually commit files

* Address review feedback

* Format

* Apply `performInvokeReturn` to forks as well

* Format

* Working on PR feedback

* Add call fail test

* Skip tests

Currently longjmp is broken (needs to be investigated), however that's
unrelated to invoke so we want to merge that in

* Format

* Fix review feedback