Make function allocation size = number of instructions
Previously, all function allocations had a constant size. However, we'd
like to be able to refer to individual instructions (e.g. for longjmp).
The new approach is to have the size of the function allocation size be
the number of instructions within the function. That way, we can refer
to individual instructions using a pointer with the appropriate offset.