Add external stackframe (#504)
This PR adds external stack frames which will be used to track return and resume values for functions called by external functions. It also adds a utility method to setup and tear down external stack frames before / after a function is called from an external method.

I didn't tie this all together yet because in the `Context` we have a vector of `StackFrame`s. However, since `ExternalStackFrame` subclasses from `StackFrame` we cannot store it in the same vector. I think the solution will be to use smartpointers in the `Context` but didn't want to add such a big change in one PR.