Add formatting targets to bazel (#512)
* Add support for a local bazelrc file

* Add top-level :format and :check-format targets

This commit defines two targets that can be used for checking formatting
and applying formatting, respectively. This means that formatting can be
checked by running

$ bazel run :check-format

and all source files can be reformatted by running

$ bazel run :format

At it's core, this works by defining a bazel aspect which aggregates all
the source files within the top-level repository which were involved in
building a target. At the end this info can then be used to generate a
script which calls clang-format on each of the relevant source files.
Actually, running the command will use a locally-built clang-format
instance from the llvm workspace so this will end up being rather slow
the first time around.

The current set of targets that will be included in formatting is
defined in the top level BUILD. Due to how the aspect implementation
works this will automatically include source files from dependencies but
it will not automatically include sources from new leaf targets. Within
the limitations of bazel this should be good enough for now.

I've also included a couple of small bazel fixes and formatting changes
that I ran into while attempting to build this.

* Skip clang-format where the two formatting tests disagree

* Add github workflow