Hide stderr output from tests that throw UnsupportedOperation exceptions
Throwing UnsupportedOperation always writes directly to stderr.
GoogleTest does not automatically catch this in the case where the test
passes. Since the output is rather noisy then this makes trying to debug
other test failures harder.

This commit addresses this by capturing stderr output to a file and then
only printing that output to stderr if the test case fails. In cases
where this doesn't work then we fail open and just allow printing to
stderr.