Only build LLVM with the X86 backend
We don't actually need any of the LLVM backends so we can save some time
by disabling as many as possible before LLVM stops compiling.
Unfortunately, I wasn't able to disable all of them but this disables
the X86 one and should cut down on a number of files that need to be
built.