Add utilities for handling function returns to `InterpreterContext` (#554)
* Add utility methods for jumping from a return/resume

* Change getCurrentInstruction to actually get the currently executing
instruction

Currently getCurrentInstruction actually returns the instruction that is
scheduled to be executed next. This isn't really useful since we usually
want to do things based on the current instruction and the next
instruction is liable to change as code executes.

This commit changes getCurrentInstruction to return the one that came
just before the one Context::current points to. It also makes sure to
take into account possible jumps to the current basic block.