Fix bug within OperationBuilder::createICmp
This overload that took LLVMValue instances was assuming that all
LLVMScalars that made up the LLVMValue were all expressions and not
pointers. This was trivially not true in a lot of cases. This commit
fixes it to use to_expr so that it properly converts pointer operands to
their integer values before building the ICmp operation.