[bazel] Make :check-format fail if formatting is invalid
Previously, this would print some output if it found formatting errors
and then exit with return code 0. Bazel would see the return code and
helpfully suppress all the output produced by the test since it was
assuming it passed. Now, the generated script will actually fail if the
formatting is not valid.