Fix `strlen` test (#596)
* Unbust `libc` linking

Previously `libc` would not be linked to `c` files, only `c++` files.
This PR fixes that

* Fix `strlen` test

This commit fixes the musl libc `strlen` implementation. The root cause
if that when `__GNUC__` is defined, the `strlen` implementation assumes
that it can read past the buffer by a bit. However, under caffeine's
memory model this is not true.

To prevent clang from defining the GNU C macro, we set
`-fgnuc-version=0` as a flag to clang.