Only link in stdlibs for appropriate languages (#638)
This commit changes how dependencies are linked in for bazel tests.
Currently, we link in libc and libc++ regardless of which language is
actually used. This is undesirable because `.ll` test files don't need
these deps and linking them in can cause builds to fail due to
redefinition of symbols.