Build all build tools in host configuration
Bazel provides the host configuration for tools so that it is not
necessary to rebuild them every time we change configuration. However,
in order to prioritize build time we've been building them using the
"exec" configuration instead. This has two main consequences:
- LLVM libaries depended on by caffeine binaries are only built once
  instead of needing to be built twice.
- We need to rebuild (or have cached) all the binaries required to build
  LLVM + clang whenever we switch compilation mode (e.g. from fastbuild
  to dbg or opt)

By switching to building all tools in the host configuration we can
resolve this issue at the expense of requiring a full clang rebuild for
everyone.

The one extra thing I needed to do was to mark `//tools/opt-plugin` with
the "manual" tag so that it is excluded from being built in the "exec"
configuration when we run `bazel build ...` (it is still built in the
"host" configuration since that is required).