Add targets to create a release package for caffeine (#670)
This PR uses `rules_pkg` to allow bazel to build a release tarball that installs all the required components to use caffeine from a cmake project. This includes:
- All visible headers from caffeine or any of its dependencies
- The caffeine binary itself
- Caffeine as a shared library which bundles all of its dependencies
- The existing `opt` plugin
- Bitcode files for caffeine-builtins, libc, and libcxx
- `caffeine-config.cmake`
- and any scripts that are meant to be used with vcpkg

In addition, I've also added an install target that allows for installing caffeine to a destination of your choice. Run `bazel run //:install -- --help` to see the flags that can be used here. Beyond that, I've also added an export step to CI so we can get a release tarball from any commit that passed CI if we so choose.

This should be enough to be able to use caffeine from cmake without needing to go through the cmake build system for caffeine at all.