Clean up unused bazel dependencies (#676)
This is a bit of an attempt to reduce some the build times for caffeine. Currently, in order to build caffeine we link in a large part of LLVM. This is slow so, naturally, we would like to avoid that.

This PR does the following:
- Remove divine as a dependency from caffeine. Last time I tried to do this we weren't sure whether it would still be needed so I haven't actually gone and deleted it from the repo just yet. However, it is now dead code.
- Remove dependencies on llvm components that aren't actually needed within the `//:caffeine` target.
- The same goes for all the tools targets as well.
- Re-add all the dependencies that weren't required for `//:caffeine` but were needed by downstream targets (e.g. `//test/unit`) as dependencies for the downstream targets.

The end result of this is that it should no longer be necessary to build all the LLVM passes in order to build caffeine. It will still be necessary to build them in the host configuration but that is a one-time that doesn't need to be redone for every configuration.