Add a test verifying that rewrites complete in time
This is a problem when we have a long chain of operations that can be
rewritten due to commutativity and associativity. In this case the
rewrites will generate exponentially many enodes as all the rewrites end
up being applied. This is suboptimal so I'm adding a test case for it so
it can be fixed.