Add a filter type for subclauses (#701)
None of the types that could potentially use this have been added
yet. However, this class was at the bottom of the hierarchy when
I was trying to split up the PR so it gets to be introduced first.  In
short, this is just an interface with some extra bits so that it remains
comparable and hashable (since that will matter in the coming PRs).