Add matcher Clauses and SubClauses (#703)
This PR introduces the `Clause` and `SubClause` types which are used to match against e-classes in the graph.

The `SubClause` type matches against an individual e-node. It can include an opcode requirement, as well as refer to other subclauses that need to match on the operands of the e-node in order for the current subclause to match. For more custom filters, it can optionally include a `SubClauseFilter` instance (added in #701) which can do custom filtering beyond that.

The `Clause` type is then a top-level matching item that will correspond to a rewrite rule. It contains the actual rewrite function (`update`), a reference to the top-level subclause that it matches against, and, optionally, a global filter that can check for non-local requirements.

In the next PR, I'll introduce the `EMatcher` class which is responsible for actually storing all of these in a way that can be used.