Add an e-graph accessor class for use in rewrite rules (#705)
This PR introduces the `GraphAccessor` class. All it does is prevent merges from happening while the rewrite rules are being applied, while recording which ones should happen so that they can all be done at once after the rewrites have occurred. See the comment on the class for more details.