Add implementation of e-matching and equality saturation (#707)
This commit is the main algorithmic part of the whole work on e-graphs and e-matching. It implements the logic for efficiently finding all queries that match within the `EMatcher` and then, for each matching query, it will apply the corresponding rewrite rule.

The algorithm implemented here is based off of the `equality_saturation` algorithm from within the [egg paper](https://arxiv.org/pdf/2004.03082.pdf). However, the details of the algorithm as implemented here end up spanning about 150 LOC. Check out `EGraphMatcher::equality_saturation` for the high level bits of code involved, and then look at the individual methods for the details of how each step is actually done.