Undo optimizations to GraphAccessor::add_merge
Since e-matching requires reference stability for both e-classes and
e-nodes it is not valid to be modifying the set of e-nodes within any
e-class in the graph within the GraphAccessor.

Unfortunately, this means that it's necessary to create a bunch of
deadweight e-classes that are then immediately merged at the end of the
matching iteration.