Implement a rewrite that implements associativity rules
The naive way to implement the associativity rule is just to blindly
apply ($op ?x ($op ?y ?z)) -> ($op ($op ?x ?y) ?z) for each operand type
that is associative until all possible trees are inserted into the
e-graph. Unfortunately, this results in exponentially many rewrites and
so ends up being completely impractical as soon as a program does
anything that generates a repeated chain of operands. (e.g. by checking
whether a multiplication would overflow).

To work around this issue, I've instead elected to pick a canonical
ordering and build an expression that results in this. I've done this by
selecting the add expression with minimal cost and then sorting the add
operands in a way that results in constant e-classes being placed
together (for constant folding) and otherwise placing similar operation
types together.