Add a non-owning erased function type
This commit introduces a new erased function type called function_view.
It is meant to be somewhat similar to std::function with the main
difference being that it doesn't actually take ownership over the
function but instead serves a function reference.

The idea is based off of the std::function_ref proposal for C++ but the
implementation, while somewhat inspired by a blog post I found on
something very similar, is mainly my own.

The code in this commit should work equally well with function pointers
as with arbitrary callable objects and the test cases included verify
that everything works as expected here.