Add support for captures to ematching subclauses
Currently, when we write a rewrite update function we need to iterate
over all the possible matches within the update function. This is
tedious and somewhat difficult to do correctly, especially as the
expression tree that is being matched gets more and more complicated.

To address this, I'm introducing the idea of a capture. This means that
the corresponding ENode can be easily accessed by the submatcher ID
without needing to manually iterate over the appropriate matches within
the update function for each and every rewrite.

The code to match and prepare the captures will be added in a future
commit.